pyproject.toml
[tool.poetry]
package-mode = false
name = "appium-android-tests"
version = "0.1.0"
description = "A sample Python project for Appium Android tests"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
pytest = "*"
allure-pytest = "*"
selene = ">=2.0.0rc4"
Appium-Python-Client = "*"

[tool.poetry.dev-dependencies]
pytest = "^7.0"